feat: update delete & remove double check modal of user group and user
