+++ apps/web/src/services/iam/pages/UserGroupMainPage.vue
@@ -7,6 +7,7 @@ import UserGroupManagementEditModal from '@/services/iam/components/UserGroupMan
 import UserGroupManagementHeader from '@/services/iam/components/UserGroupManagementHeader.vue';
 import UserGroupManagementTab from '@/services/iam/components/UserGroupManagementTab.vue';
 import UserGroupManagementTable from '@/services/iam/components/UserGroupManagementTable.vue';
+import UserPerGroupRemoveDoubleCheckModal from '@/services/iam/components/UserPerGroupRemoveDoubleCheckModal.vue';
 import { useUserGroupPageStore } from '@/services/iam/store/user-group-page-store';
 
 const userGroupPageStore = useUserGroupPageStore();
@@ -35,6 +36,7 @@ const refreshUserGroupList = () => {
         <user-group-management-edit-modal @confirm="refreshUserGroupList" />
         <user-group-management-add-users-modal />
         <user-group-delete-double-check-modal />
+        <user-per-group-remove-double-check-modal />
     </section>
 </template>
